service/requirements.txt
Flask-Cors==1.0
Flask-Restless
Flask-Login
SQLAlchemy==0.8.4
